Moved ReferenceManagerPortlet to proper package
and created separate BibliographyManagerPortlet permissions.
