+++ reference-manager/reference-manager-api/src/main/java/ch/inofix/referencemanager/constants/PortletKeys.java
@@ -4,13 +4,13 @@ package ch.inofix.referencemanager.constants;
  * 
  * @author Christian Berndt
  * @created 2016-11-18 21:56
- * @modified 2016-12-01 02:04
- * @version 1.0.1
+ * @modified 2016-12-02 17:44
+ * @version 1.0.2
  * 
  */
 public class PortletKeys {
 
     public static final String BIBLIOGRAPHY_MANAGER = "ch_inofix_referencemanager_web_internal_portlet_BibliographyManagerPortlet";
-    public static final String REFERENCE_MANAGER = "ch_inofix_referencemanager_web_ReferenceManagerPortlet";
+    public static final String REFERENCE_MANAGER = "ch_inofix_referencemanager_web_internal_portlet_ReferenceManagerPortlet";
     public static final String YOUR_BIBLIOGRAPHIES = "ch_inofix_referencemanager_web_internal_portlet_YourBibliographiesPortlet";
 }
