+++ reference-manager/reference-manager-service/src/main/resources/META-INF/resource-actions/default.xml
@@ -5,9 +5,9 @@
 
 <resource-action-mapping>
     <model-resource>
-        <model-name>ch.inofix.referencemanager</model-name>
+        <model-name>ch.inofix.referencemanager.bibliography</model-name>
         <portlet-ref>
-            <portlet-name>ch_inofix_referencemanager_web_ReferenceManagerPortlet</portlet-name>
+            <portlet-name>ch_inofix_referencemanager_web_internal_portlet_BibliographyManagerPortlet</portlet-name>
         </portlet-ref>
         <root>true</root>
         <weight>1</weight>
@@ -15,7 +15,6 @@
             <supports>
                 <action-key>ADD_BIBLIOGRAPHY</action-key>
                 <action-key>ADD_REFERENCE</action-key>
-                <action-key>DELETE_GROUP_REFERENCES</action-key>
                 <action-key>IMPORT_BIBLIOGRAPHY</action-key>
                 <action-key>IMPORT_REFERENCES</action-key>
                 <action-key>PERMISSIONS</action-key>
@@ -35,7 +34,6 @@
             <guest-unsupported>
                 <action-key>ADD_BIBLIOGRAPHY</action-key>
                 <action-key>ADD_REFERENCE</action-key>
-                <action-key>DELETE_GROUP_REFERENCES</action-key>
                 <action-key>IMPORT_BIBLIOGRAPHY</action-key>
                 <action-key>IMPORT_REFERENCES</action-key>
                 <action-key>PERMISSIONS</action-key>
@@ -46,7 +44,7 @@
     <model-resource>
         <model-name>ch.inofix.referencemanager.model.Bibliography</model-name>
         <portlet-ref>
-            <portlet-name>ch_inofix_referencemanager_web_ReferenceManagerPortlet</portlet-name>
+            <portlet-name>ch_inofix_referencemanager_web_internal_portlet_BibliographyManagerPortlet</portlet-name>
         </portlet-ref>
         <weight>3</weight>
         <permissions>
@@ -73,9 +71,42 @@
         </permissions>
     </model-resource>
     <model-resource>
+        <model-name>ch.inofix.referencemanager.reference</model-name>
+        <portlet-ref>
+            <portlet-name>ch_inofix_referencemanager_web_internal_portlet_ReferenceManagerPortlet</portlet-name>
+        </portlet-ref>
+        <root>true</root>
+        <weight>1</weight>
+        <permissions>
+            <supports>
+                <action-key>ADD_REFERENCE</action-key>
+                <action-key>DELETE_GROUP_REFERENCES</action-key>
+                <action-key>IMPORT_REFERENCES</action-key>
+                <action-key>PERMISSIONS</action-key>
+                <action-key>SUBSCRIBE</action-key>
+                <action-key>VIEW</action-key>
+            </supports>
+            <site-member-defaults>
+                <action-key>ADD_REFERENCE</action-key>
+                <action-key>SUBSCRIBE</action-key>
+                <action-key>VIEW</action-key>
+            </site-member-defaults>
+            <guest-defaults>
+                <action-key>VIEW</action-key>
+            </guest-defaults>
+            <guest-unsupported>
+                <action-key>ADD_REFERENCE</action-key>
+                <action-key>DELETE_GROUP_REFERENCES</action-key>
+                <action-key>IMPORT_REFERENCES</action-key>
+                <action-key>PERMISSIONS</action-key>
+                <action-key>SUBSCRIBE</action-key>
+            </guest-unsupported>
+        </permissions>
+    </model-resource>    
+    <model-resource>
         <model-name>ch.inofix.referencemanager.model.Reference</model-name>
         <portlet-ref>
-            <portlet-name>ch_inofix_referencemanager_web_ReferenceManagerPortlet</portlet-name>
+            <portlet-name>ch_inofix_referencemanager_web_internal_portlet_ReferenceManagerPortlet</portlet-name>
         </portlet-ref>
         <weight>3</weight>
         <permissions>
