+++ reference-manager/reference-manager-web/src/main/resources/META-INF/resource-actions/default.xml
@@ -3,7 +3,26 @@
 
 <resource-action-mapping>
     <portlet-resource>
-        <portlet-name>ch_inofix_referencemanager_web_ReferenceManagerPortlet</portlet-name>
+        <portlet-name>ch_inofix_referencemanager_web_internal_portlet_BibliographyManagerPortlet</portlet-name>
+        <permissions>
+            <supports>
+                <action-key>ADD_TO_PAGE</action-key>
+                <action-key>CONFIGURATION</action-key>
+                <action-key>VIEW</action-key>
+            </supports>
+            <site-member-defaults>
+                <action-key>VIEW</action-key>
+            </site-member-defaults>
+            <guest-defaults>
+                <action-key>VIEW</action-key>
+            </guest-defaults>
+            <guest-unsupported>
+                <action-key>CONFIGURATION</action-key>
+            </guest-unsupported>
+        </permissions>
+    </portlet-resource>
+    <portlet-resource>
+        <portlet-name>ch_inofix_referencemanager_web_internal_portlet_ReferenceManagerPortlet</portlet-name>
         <permissions>
             <supports>
                 <action-key>ADD_TO_PAGE</action-key>
