+++ reference-manager/reference-manager-web/src/main/resources/content/Language.properties
@@ -13,13 +13,13 @@ delete-all-references=Delete all References
 import-export=Import / Export
 javax.portlet.title.ch_inofix_referencemanager_web_internal_portlet_BibliographyManagerPortlet=Bibliography Manager
 javax.portlet.title.ch_inofix_referencemanager_web_internal_portlet_YourBibliographiesPortlet=Your Bibliographies
-javax.portlet.title.ch_inofix_referencemanager_web_ReferenceManagerPortlet=Reference Manager
+javax.portlet.title.ch_inofix_referencemanager_web_internal_portlet_ReferenceManagerPortlet=Reference Manager
 javax.portlet.display-name.ch_inofix_referencemanager_web_internal_portlet_BibliographyManagerPortlet=Bibliography Manager
 javax.portlet.display-name.ch_inofix_referencemanager_web_internal_portlet_YourBibliographiesPortlet=Your Bibliographies
-javax.portlet.display-name.ch_inofix_referencemanager_web_ReferenceManagerPortlet=Reference Manager
+javax.portlet.display-name.ch_inofix_referencemanager_web_internal_portlet_ReferenceManagerPortlet=Reference Manager
 javax.portlet.short-title.ch_inofix_referencemanager_web_internal_portlet_BibliographyManagerPortlet=Bibliography Manager
 javax.portlet.short-title.ch_inofix_referencemanager_web_internal_portlet_YourBibliographiesPortlet=Your Bibliographies
-javax.portlet.short-title.ch_inofix_referencemanager_web_ReferenceManagerPortlet=Reference Manager
+javax.portlet.short-title.ch_inofix_referencemanager_web_internal_portlet_ReferenceManagerPortlet=Reference Manager
 model.resource.ch.inofix.referencemanager.model.Bibliography=BibTeX Bibliography
 model.resource.ch.inofix.referencemanager.model.Reference=BibTeX Reference
 new-bibliography=New Bibliography
